added vslam capability

+++ simulation/ws_warehouse_drone/install/px4_msgs/local/lib/python3.10/dist-packages/px4_msgs/msg/_vehicle_command_s.c
@@ -164,7 +164,7 @@ bool px4_msgs__msg__vehicle_command__convert_from_py(PyObject * _pymsg, void * _
       return false;
     }
     assert(PyLong_Check(field));
-    ros_message->source_component = (uint8_t)PyLong_AsUnsignedLong(field);
+    ros_message->source_component = (uint16_t)PyLong_AsUnsignedLong(field);
     Py_DECREF(field);
   }
   {  // confirmation
